refactor : custom annotation 적용
